Refactor authentication flow and layout structure for login and dashboard pages
