+++ src/app/api/transactions/page.tsx
@@ -1,4 +1,3 @@
-// app/dashboard/transactions/page.tsx
 "use client";
 
 import { useEffect, useState } from "react";
